FlaskApps/requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysqlclient==2.2.7
pycryptodome==3.21.0
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2024.2
Werkzeug==3.1.3
